Golden.cpp/run all students
This code mostly works for compiling and testing each students source
code, it doesn't do anything to do with the log files though (wasn't
sure if that was me or Ben).

There's a slight problem when running the very first student but I think
that has something to do with changing back to the correct directory so
that should be an easy fix. Also, fully implemented to "golden.cpp"
stuff so that it's hopefully compatible with James' code.