Mostly Done
This documentation is mostly complete. I will change it, if need be,
after we all get a chance to read through.

I have upgraded their documentation for the design of the algorithm,
deployment of the program, development (tools, ide,etc.) , and the user
documentation specifying how to use the program with the test generation
features.

I know I am missing the test header and plan to work on that tonight.